{"schema":"decodex/commit/1","summary":"Modularize runtime persistence and recovery boundaries","authority":"manual"}
