+++ apps/decodex/src/orchestrator/execution/app_server/continuation.rs
@@ -1,0 +1,19 @@
+use crate::orchestrator::{
+	self, IssueRunPlan, ReviewHandoffContext, ServiceConfig, WorkflowDocument,
+};
+
+pub(super) fn build_issue_run_continuation_user_input(
+	project: &ServiceConfig,
+	workflow: &WorkflowDocument,
+	issue_run: &IssueRunPlan,
+	review_context: &ReviewHandoffContext,
+) -> String {
+	orchestrator::build_continuation_user_input(
+		&issue_run.issue,
+		workflow,
+		issue_run.dispatch_mode,
+		review_context.recorded_pr_url.as_deref(),
+		workflow.frontmatter().tracker().success_state(),
+		project.codex().review_level(),
+	)
+}
