+++ apps/decodex/src/orchestrator/execution/app_server/failure.rs
@@ -1,0 +1,65 @@
+use color_eyre::Report;
+
+use crate::orchestrator::{
+	self, IssueRunPlan, IssueTracker, Result, RunSummary, ServiceConfig, StateStore,
+	TrackerToolBridge, WorkflowDocument, execution::completion,
+};
+
+pub(super) fn maybe_finalize_after_terminalized_app_server_failure<T>(
+	tracker: &T,
+	project: &ServiceConfig,
+	workflow: &WorkflowDocument,
+	state_store: &StateStore,
+	issue_run: &IssueRunPlan,
+	tracker_tool_bridge: &TrackerToolBridge<'_>,
+	error: &Report,
+) -> Result<Option<RunSummary>>
+where
+	T: IssueTracker,
+{
+	let Some(disposition) = tracker_tool_bridge.finalized_completion_disposition()? else {
+		return Ok(None);
+	};
+
+	state_store.append_private_execution_event(
+		project.service_id(),
+		&issue_run.issue.id,
+		&issue_run.run_id,
+		issue_run.attempt_number,
+		"terminal_finalize_app_server_failure_recovery",
+		serde_json::json!({
+			"path": disposition.as_str(),
+			"source_error": error.to_string(),
+			"recovery": "apply_terminal_completion_writeback",
+		}),
+	)?;
+
+	tracing::warn!(
+		project_id = project.service_id(),
+		issue_id = issue_run.issue.id,
+		issue = issue_run.issue.identifier,
+		run_id = issue_run.run_id,
+		attempt = issue_run.attempt_number,
+		path = disposition.as_str(),
+		error = %error,
+		"App-server run failed after terminal finalize; applying terminal completion writeback."
+	);
+
+	completion::apply_run_completion_disposition(
+		tracker,
+		project,
+		workflow,
+		state_store,
+		issue_run,
+		tracker_tool_bridge,
+	)?;
+
+	state_store.record_run_attempt(
+		&issue_run.run_id,
+		&issue_run.issue.id,
+		issue_run.attempt_number,
+		"succeeded",
+	)?;
+
+	Ok(Some(orchestrator::run_summary_from_issue_run(project.service_id(), issue_run)))
+}
