+++ apps/decodex/src/recovery/review_handoff_diagnosis.rs
@@ -2,20 +2,16 @@
 
 mod actions;
 mod binding;
+mod issue_refresh;
+mod stale_residue;
 
 pub(super) use crate::recovery::review_handoff_diagnosis::binding::{
 	HandoffDiagnosticRequest, diagnostic_binding,
 };
 
-use std::collections::{HashMap, HashSet};
-
 use crate::{
-	orchestrator,
 	prelude::{Result, eyre},
-	recovery::{
-		self, REVIEW_HANDOFF_STALE_TERMINAL_RESIDUE_CLASSIFICATION, context::RecoveryContext,
-		git_worktree, reports::ReviewHandoffDiagnostic,
-	},
+	recovery::{self, context::RecoveryContext, git_worktree, reports::ReviewHandoffDiagnostic},
 	state::WorktreeMapping,
 	tracker::{self, IssueTracker, TrackerIssue},
 };
@@ -37,14 +33,16 @@ where
 	let mut diagnostics = Vec::new();
 
 	for worktree in context.state_store.list_worktrees(context.config.service_id())? {
-		if retained_review_worktree_is_stale_terminal_residue(context, &worktree)? {
-			diagnostics.push(stale_terminal_residue_review_handoff_diagnostic(context, &worktree));
+		if stale_residue::retained_review_worktree_is_stale_terminal_residue(context, &worktree)? {
+			diagnostics.push(stale_residue::stale_terminal_residue_review_handoff_diagnostic(
+				context, &worktree,
+			));
 		} else {
 			worktrees.push(worktree);
 		}
 	}
 
-	let issues_by_id = refresh_retained_review_worktree_issues(tracker, &worktrees)?;
+	let issues_by_id = issue_refresh::refresh_retained_review_worktree_issues(tracker, &worktrees)?;
 	let tracker_policy = context.workflow.frontmatter().tracker();
 	let success_state = tracker_policy.success_state();
 	let in_progress_state = tracker_policy.in_progress_state();
@@ -83,8 +81,12 @@ pub(super) fn diagnose_issue_with_tracker<T>(
 where
 	T: IssueTracker,
 {
-	if let Some(worktree) = stale_terminal_residue_worktree_for_issue(context, issue_identifier)? {
-		return Ok(stale_terminal_residue_review_handoff_diagnostic(context, &worktree));
+	if let Some(worktree) =
+		stale_residue::stale_terminal_residue_worktree_for_issue(context, issue_identifier)?
+	{
+		return Ok(stale_residue::stale_terminal_residue_review_handoff_diagnostic(
+			context, &worktree,
+		));
 	}
 
 	let issue = recovery::load_issue_by_identifier(tracker, issue_identifier)?;
@@ -93,93 +95,6 @@ where
 	})?;
 
 	diagnose_issue_worktree(context, issue, worktree)
-}
-
-fn retained_review_worktree_is_stale_terminal_residue(
-	context: &RecoveryContext,
-	worktree: &WorktreeMapping,
-) -> Result<bool> {
-	let active_issue_ids = context
-		.state_store
-		.list_active_shared_leases(context.config.service_id())?
-		.into_iter()
-		.map(|lease| lease.issue_id().to_owned())
-		.collect::<HashSet<_>>();
-
-	orchestrator::worktree_mapping_is_stale_terminal_local_residue(
-		&context.config,
-		&context.state_store,
-		worktree,
-		&active_issue_ids,
-	)
-}
-
-fn stale_terminal_residue_review_handoff_diagnostic(
-	context: &RecoveryContext,
-	worktree: &WorktreeMapping,
-) -> ReviewHandoffDiagnostic {
-	ReviewHandoffDiagnostic {
-		project_id: context.config.service_id().to_owned(),
-		issue_id: worktree.issue_id().to_owned(),
-		issue_identifier: worktree.issue_id().to_owned(),
-		issue_state: String::from("local_terminal_residue"),
-		classification: String::from(REVIEW_HANDOFF_STALE_TERMINAL_RESIDUE_CLASSIFICATION),
-		reason: String::from(
-			"terminal_unleased_runtime_recorded_identifier_mapping_with_missing_path",
-		),
-		branch_name: worktree.branch_name().to_owned(),
-		worktree_path: worktree.worktree_path().display().to_string(),
-		local_branch_name: None,
-		local_head_oid: None,
-		worktree_clean: None,
-		existing_pr_url: None,
-		existing_lifecycle_handoff_head_oid: None,
-		existing_lifecycle_phase_head_oid: None,
-		pr_base_ref: None,
-		pr_head_oid: None,
-		pr_read_error: None,
-		mismatched_field: None,
-		active_label_present: None,
-		next_action: String::from(
-			"No review-handoff recovery action is required; project reconciliation clears this stale local mapping before tracker refresh.",
-		),
-	}
-}
-
-fn refresh_retained_review_worktree_issues<T>(
-	tracker: &T,
-	worktrees: &[WorktreeMapping],
-) -> Result<HashMap<String, TrackerIssue>>
-where
-	T: IssueTracker,
-{
-	if worktrees.is_empty() {
-		return Ok(HashMap::new());
-	}
-
-	let issue_ids =
-		worktrees.iter().map(|worktree| worktree.issue_id().to_owned()).collect::<Vec<_>>();
-
-	Ok(tracker
-		.refresh_issues(&issue_ids)?
-		.into_iter()
-		.map(|issue| (issue.id.clone(), issue))
-		.collect())
-}
-
-fn stale_terminal_residue_worktree_for_issue(
-	context: &RecoveryContext,
-	issue_identifier: &str,
-) -> Result<Option<WorktreeMapping>> {
-	let Some(worktree) = context.state_store.worktree_for_issue(issue_identifier)? else {
-		return Ok(None);
-	};
-
-	if retained_review_worktree_is_stale_terminal_residue(context, &worktree)? {
-		Ok(Some(worktree))
-	} else {
-		Ok(None)
-	}
 }
 
 fn diagnose_issue_worktree(
