+++ apps/decodex/src/state/sqlite_store/mutations/runs/attempts.rs
@@ -1,0 +1,25 @@
+use crate::state::sqlite_store::mutations::{self, Result, RunAttemptRecord, SqliteStateStore};
+
+impl SqliteStateStore {
+	pub(in crate::state) fn upsert_run_attempt(&self, attempt: &RunAttemptRecord) -> Result<()> {
+		self.connection.execute(
+			"INSERT OR REPLACE INTO run_attempts (
+					run_id, project_id, issue_id, attempt_number, status, thread_id, turn_id,
+					updated_at, updated_at_unix
+				) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
+			mutations::params![
+				&attempt.run_id,
+				attempt.project_id.as_deref(),
+				&attempt.issue_id,
+				attempt.attempt_number,
+				&attempt.status,
+				attempt.thread_id.as_deref(),
+				attempt.turn_id.as_deref(),
+				&attempt.updated_at,
+				attempt.updated_at_unix,
+			],
+		)?;
+
+		Ok(())
+	}
+}
