+++ apps/decodex/src/state/sqlite_store/persist/base/runs.rs
@@ -1,0 +1,93 @@
+use crate::state::sqlite_store::persist::{
+	self, ChildAgentActivitySummary, Result, StateData, Transaction,
+};
+
+pub(in crate::state::sqlite_store) fn persist_run_attempts(
+	transaction: &Transaction<'_>,
+	state: &StateData,
+) -> Result<()> {
+	for attempt in state.run_attempts.values() {
+		transaction.execute(
+			"INSERT OR REPLACE INTO run_attempts (
+					run_id, project_id, issue_id, attempt_number, status, thread_id, turn_id,
+					updated_at, updated_at_unix
+				) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
+			persist::params![
+				&attempt.run_id,
+				attempt.project_id.as_deref(),
+				&attempt.issue_id,
+				attempt.attempt_number,
+				&attempt.status,
+				attempt.thread_id.as_deref(),
+				attempt.turn_id.as_deref(),
+				&attempt.updated_at,
+				attempt.updated_at_unix,
+			],
+		)?;
+	}
+
+	Ok(())
+}
+
+pub(in crate::state::sqlite_store) fn persist_run_control_channels(
+	transaction: &Transaction<'_>,
+	state: &StateData,
+) -> Result<()> {
+	for channel in state.control_channels.values() {
+		transaction.execute(
+			"INSERT OR REPLACE INTO run_control_channels (
+					run_id, project_id, issue_id, attempt_number, transport, channel_path, status,
+					published_at, published_at_unix, updated_at, updated_at_unix
+				) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11)",
+			persist::params![
+				&channel.run_id,
+				&channel.project_id,
+				&channel.issue_id,
+				channel.attempt_number,
+				&channel.transport,
+				channel.channel_path.to_string_lossy().as_ref(),
+				&channel.status,
+				&channel.published_at,
+				channel.published_at_unix,
+				&channel.updated_at,
+				channel.updated_at_unix,
+			],
+		)?;
+	}
+
+	Ok(())
+}
+
+pub(in crate::state::sqlite_store) fn persist_run_activity_summaries(
+	transaction: &Transaction<'_>,
+	state: &StateData,
+) -> Result<()> {
+	for summary in state.run_activity_summaries.values() {
+		let child_agent_activity_json = summary
+			.child_agent_activity
+			.as_ref()
+			.cloned()
+			.map(ChildAgentActivitySummary::sealed_durable)
+			.map(|summary| serde_json::to_string(&summary))
+			.transpose()?;
+		let protocol_activity_json =
+			summary.protocol_activity.as_ref().map(serde_json::to_string).transpose()?;
+
+		transaction.execute(
+			"INSERT OR REPLACE INTO run_activity_summaries (
+					run_id, attempt_number, child_agent_activity_json, protocol_activity_json,
+					updated_at, updated_at_unix
+				) VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
+			persist::params![
+				&summary.run_id,
+				summary.attempt_number,
+				child_agent_activity_json.as_deref(),
+				protocol_activity_json.as_deref(),
+				&summary.updated_at,
+				summary.updated_at_unix,
+			],
+		)?;
+	}
+
+	Ok(())
+}
