+++ apps/decodex/src/state/sqlite_store/queries/runs/lookup.rs
@@ -1,0 +1,71 @@
+use crate::state::sqlite_store::{
+	SqliteStateStore,
+	queries::{self, RunAttemptRecord, run_attempt_record_from_row},
+};
+
+impl SqliteStateStore {
+	#[cfg(test)]
+	pub(in crate::state) fn run_attempt_for_issue_attempt(
+		&self,
+		issue_id: &str,
+		attempt_number: i64,
+	) -> queries::Result<Option<RunAttemptRecord>> {
+		let mut statement = self.connection.prepare(
+			"SELECT run_id, project_id, issue_id, attempt_number, status, thread_id, turn_id, \
+			 updated_at, updated_at_unix FROM run_attempts \
+			 WHERE issue_id = ?1 AND attempt_number = ?2 \
+			 ORDER BY updated_at_unix DESC, run_id DESC \
+			 LIMIT 1",
+		)?;
+		let mut rows = statement.query(queries::params![issue_id, attempt_number])?;
+
+		Ok(rows.next()?.map(run_attempt_record_from_row).transpose()?)
+	}
+
+	pub(in crate::state) fn latest_run_attempt_for_issue(
+		&self,
+		issue_id: &str,
+	) -> queries::Result<Option<RunAttemptRecord>> {
+		let mut statement = self.connection.prepare(
+			"SELECT run_id, project_id, issue_id, attempt_number, status, thread_id, turn_id, \
+			 updated_at, updated_at_unix FROM run_attempts \
+			 WHERE issue_id = ?1 \
+			 ORDER BY attempt_number DESC, updated_at_unix DESC, run_id DESC \
+			 LIMIT 1",
+		)?;
+		let mut rows = statement.query(queries::params![issue_id])?;
+
+		Ok(rows.next()?.map(run_attempt_record_from_row).transpose()?)
+	}
+
+	pub(in crate::state) fn list_run_attempts_for_issue(
+		&self,
+		issue_id: &str,
+	) -> queries::Result<Vec<RunAttemptRecord>> {
+		let mut statement = self.connection.prepare(
+			"SELECT run_id, project_id, issue_id, attempt_number, status, thread_id, turn_id, \
+			 updated_at, updated_at_unix FROM run_attempts \
+			 WHERE issue_id = ?1 \
+			 ORDER BY attempt_number ASC, run_id ASC",
+		)?;
+		let rows = statement.query_map(queries::params![issue_id], run_attempt_record_from_row)?;
+
+		rows.collect::<rusqlite::Result<Vec<_>>>().map_err(Into::into)
+	}
+
+	pub(in crate::state) fn list_run_attempts_for_project(
+		&self,
+		project_id: &str,
+	) -> queries::Result<Vec<RunAttemptRecord>> {
+		let mut statement = self.connection.prepare(
+			"SELECT run_id, project_id, issue_id, attempt_number, status, thread_id, turn_id, \
+			 updated_at, updated_at_unix FROM run_attempts \
+			 WHERE project_id = ?1 \
+			 ORDER BY updated_at_unix DESC, run_id ASC",
+		)?;
+		let rows =
+			statement.query_map(queries::params![project_id], run_attempt_record_from_row)?;
+
+		rows.collect::<rusqlite::Result<Vec<_>>>().map_err(Into::into)
+	}
+}
